doc string fo solver
